# APyFixed: specialize arithmetic when result fits single limb