### Benchmark
```python
from apytypes import APyFixed
from time import time

a = APyFixed(123456, bits=24, int_bits=12)
b = APyFixed(654321, bits=25, int_bits=12)

t1 = time()
for i in range(1000000):
    c = a + b
t2 = time()
print(f"Add: {(t2 - t1)*1e3} ms")

t1 = time()
for i in range(1000000):
    c = a - b
t2 = time()
print(f"Sub: {(t2 - t1)*1e3} ms")

t1 = time()
for i in range(1000000):
    c = a * b
t2 = time()
print(f"Mul: {(t2 - t1)*1e3} ms")
```

### Before commit
```
Add: 209.1515064239502 ms
Sub: 210.9079360961914 ms
Mul: 237.6255989074707 ms
```

### After commit
```
Add: 178.8461208343506 ms
Sub: 179.54015731811523 ms
Mul: 167.3893928527832 ms
```